update: abc226a

+++ cpp/abc226/a/main.cpp
@@ -4,5 +4,5 @@ using namespace std;
 int main() {
     double x;
     cin >> x;
-    cout << round(x) << endl;
+    cout << (int)round(x) << endl;
 }
